+++ src/ui/tag_manager.py
@@ -11,12 +11,10 @@ class TagManager:
         self.page = page
         self.store = store
         self.on_changed = on_changed
-        
-        # UI components
+
         self.menu = ContextMenu()
         self.tag_dialog = TagDialog(store=store, on_changed=on_changed)
-        
-        # Dialogs
+
         self.rename_field = ft.TextField(label="New tag name", expand=True)
         self.rename_dialog = ft.AlertDialog(
             modal=True,